Added a simple mail handler.
It lets users add notes to an existing issue by replying to the initial notification email.
Permissions are checked in the same way as in the application (the user is identified by its mail address).

Information about configuring the application so that it receives emails can be found here:
http://wiki.rubyonrails.com/rails/pages/HowToReceiveEmailsWithActionMailer

RedMine mail hander is MailHandler#receive

git-svn-id: http://redmine.rubyforge.org/svn@523 e93f8b46-1217-0410-a6f0-8f06a7374b81